+++ lib/pubup.dart
@@ -13,6 +13,7 @@ library;
 
 export 'src/candidate_collector.dart'
     show CandidateUpdate, CollectionReport, CollectionResult, collectCandidates;
+export 'src/commands/self_update.dart' show runSelfUpdate;
 export 'src/outdated_runner.dart' show OutdatedPackage, getOutdatedPackages;
 export 'src/pubspec_parser.dart'
     show
@@ -21,5 +22,13 @@ export 'src/pubspec_parser.dart'
         isFlutterPackage,
         parseDependencyEntries;
 export 'src/reporter.dart' show PackageReport, printReport;
+export 'src/update_checker.dart'
+    show
+        checkForUpdate,
+        disableUpdateCheckEnv,
+        isUpdateCheckDisabled,
+        pubupPackageName,
+        resolveUpdateCacheDir;
 export 'src/updater.dart' show runUpdatesForPackage;
+export 'src/version.dart' show packageVersion;
 export 'src/workspace_discovery.dart' show discoverWorkspaceDirs, filterTargets;